task 8/requirements.txt
fastapi
uvicorn
sqlalchemy
pydantic
grpcio
grpcio-tools
